Initial commit for society project

+++ src/app/admin/admin-about/admin-about.component.ts
@@ -13,12 +13,14 @@ export class AdminAboutComponent {
 
 constructor(private router: Router) {}
 
-
   menuOpen = false;
   servicesOpen = false;
   residentsOpen = false;
   securityOpen = false;
   visitorsOpen = false;
+  eventOpen = false;
+  maintenanceOpen = false;
+  noticeOpen = false;
 
   toggleMenu() {
     this.menuOpen = !this.menuOpen;
@@ -40,17 +42,37 @@ constructor(private router: Router) {}
   this.visitorsOpen = !this.visitorsOpen;
   }
 
+  toggleEvent(){
+    this.eventOpen = !this.eventOpen;
+  }
+
+  toggleMaintenance(){
+    this.maintenanceOpen = !this.maintenanceOpen;
+  }
+
+  toggleNotice(){
+    this.noticeOpen = !this.noticeOpen;
+  }
+
 logout() {
+    Swal.fire({
+      title: 'Logout?',
+      text: 'Are you sure you want to logout?',
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonText: 'Yes, Logout'
+    }).then((result) => {
+      if (result.isConfirmed) {
         localStorage.clear();
         Swal.fire({
           icon: 'success',
           title: 'Logged Out',
-          text: 'You have been logged out successfully.',
-          timer: 1500,
+          timer: 1200,
           showConfirmButton: false
         }).then(() => {
           this.router.navigate(['/admin-login']);
         });
       }
-
+    });
+  }
 }
